backend/requirements.txt
# Core dependencies
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
alembic==1.13.1
httpx==0.25.2

# MSP-specific dependencies
pyconnectwise==0.6.2
boto3==1.34.82
celery==5.3.6
redis==5.0.4
hvac==1.2.1  # HashiCorp Vault client

# Keep.dev dependencies
# These will be installed from the Keep repository
# keep-core
# keep-providers

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5

# Development
black==24.2.0
isort==5.13.2
mypy==1.9.0
